fixes and improvements erc token
